// types.ts
export type AdGroupCriterionResource = {
  adGroupResource: string;
  AdGroupCriterionResource: string;
  criterion_id: number;
  listing_group: {
    parent_ad_group_criterion: string | null;
    case_value: {
      product_brand?: { value: string | null };
      //product_condition?: { value: string | null };
      product_type?: { value: string | null };
      product_item_id?: { value: string | null };
    } | null;
    path: {
      dimensions: Array<{
        product_brand?: { value: string };
        product_type?: { level: string; value: string };
        product_item_id?: { value: string };
      }>;
    };
    type: number;
  };
  negative: boolean;
  type: number;
  status: number;
  resource_name: string;
  clicks: number;
  conversions: number;
  cost_micros: number;
  ctr: number | null;
  conversions_value: number;
  impressions: number;
  cpc_bid_micros: number;
  average_cpc: number | null;
  idfromproduct_group_viewwheread_group_criterion: string | null;
  children: AdGroupCriterionResource[];
  level: number;
};

type ProductGroup = AdGroupCriterionResource[];

//productPartition types
export type CriterionData = {
  listingGroupType: number;
  parentId: string; // id extracted on server-side
  status: number;
  negative: boolean;
  brand: string[];
  productType: string[];
  productId: string[];
  partitionId: number;
  adGroupId: string;
};